requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
httpx==0.27.0
pydantic==1.10.14
python-dotenv==1.0.1